Custom Loader
A loader is a software component or program that is responsible for loading and preparing other software or data for execution or use by a computer system. Loaders are essential in the process of running programs and managing resources within a computing environment. 
The specific functionality of a loader depends on its purpose and the context in which it operates. Loaders are essential components of computer systems because they facilitate the execution of programs and the management of resources required for those programs to run.